SH-165 Move mediator's API related classes to mediator/api directory
