+++ central_unit/src/mediator/api/rf_api.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "rf_types.h"
+#include "../rf_types.h"
 #include "api.h"
 #include "api_client.h"
-#include "../core/api/internal_api.h"
+#include "../../core/api/internal_api.h"
 
 
 namespace sa = SmartHome::API;
